- adding simple falling block game - changing int coorinates to float in AbstractArea

+++ SoundBoard/src/SoundArea.cpp
@@ -1,6 +1,6 @@
 #include "AbstractArea.h"
 
-SoundArea::SoundArea(unsigned x, unsigned y, unsigned width, unsigned height, sf::Color color, Instrmnt *instrument, StkFloat baseFreq)
+SoundArea::SoundArea(float x, float y, float width, float height, sf::Color color, Instrmnt *instrument, StkFloat baseFreq)
 	: AbstractArea(x, y, width, height, color), baseFreq_(baseFreq)
 {
 	soundMaker_.setInstrument(instrument);
